requirements.txt
fastapi
uvicorn
sentence-transformers
elasticsearch
pandas
scikit-learn
python-multipart
tortoise-orm
pydantic
pydantic_settings